+++ dev_scripts/debugging_scripts/wrappers/watermark_test.py
@@ -1,0 +1,10 @@
+from aa_utilities.wrappers import watermark as aa_watermark
+
+output_str = aa_watermark(
+    author="Amin Allahyar",
+    email='Amin.Allahyar@astrazeneca.com',
+    timezone='Europe/Stockholm',
+    namespace=globals(),
+)
+
+print(output_str)